+++ app/src/main/java/com/chetocoders/chetogames/data/server/dto/GamemodeDTO.kt
@@ -10,6 +10,6 @@ data class GamemodeDTOResult(
 
 @Parcelize
 data class GamemodeDTO (
-    @SerializedName("id") val id: Int?,
+    @SerializedName("id") val id: Long?,
     @SerializedName("name") val name: String?
 ) : Parcelable
